# phenix/src/go/web/public/docs/openapi.yml
openapi: 3.0.0
info:
  description: gophenix API
  version: 1.0.0
  title: gophenix
  termsOfService: http://swagger.io/terms/
  contact:
    email: emulytics@sandia.gov
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/signup":
    post:
      tags:
        - Users
      summary: Signup as a new user
      description: ""
      operationId: postSignup
      requestBody:
        description: user details
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credentials"
  "/login":
    get:
      tags:
        - Users
      summary: Login a user
      description: ""
      operationId: getLogin
      parameters:
        - in: query
          name: user
          schema:
            type: string
          description: login username
        - in: query
          name: pass
          schema:
            type: string
          description: login password
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credentials"
    post:
      tags:
        - Users
      summary: Login a user
      description: ""
      operationId: postLogin
      requestBody:
        description: login credentials
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                pass:
                  type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credentials"
...
servers:
  - url: http://phenix.web/api/v1
components:
  schemas:
...
    Credentials:
      type: object
      properties:
        username:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
        token:
          type: string